[Feat] Domain Shared sub 모듈 생성, Feature에 있는 UseCase 분리
