+++ DependencyPlugin/ProjectDescriptionHelpers/Modules.swift
@@ -44,6 +44,7 @@ public extension ModulePath {
 
 public extension ModulePath {
     enum Domain: String, CaseIterable {
+        case Shared
         case OAuth
         public static let name: String = "Domain"
