+++ Projects/Domain/User/Project.swift
@@ -16,6 +16,8 @@ let project = Project.makeModule(
             interface: .User,
             factory: .init(
                 dependencies: [
+                    .domain(interface: .Shared),
+                    .domain(implements: .Shared),
                     .core
                 ]
             )
